requirements.txt
altair==5.3.0
appdirs==1.4.4
arrow==1.3.0
attrs==23.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colored==2.2.4
data-printer==0.0.8
flexcache==0.3
flexparser==0.3.1
folium==0.17.0
geopandas==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
Pint==0.24
polyline==2.0.2
protobuf==5.27.1
pyarrow==16.1.0
pydantic==1.10.17
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
stravalib==1.7
streamlit==1.36.0
streamlit_folium==0.20.1
tenacity==8.4.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
xyzservices==2024.6.0
